Conflicts:
	common/com.raytheon.uf.common.dataplugin.maps/META-INF/MANIFEST.MF
	common/com.raytheon.uf.common.python.concurrent/META-INF/MANIFEST.MF
	edex/com.raytheon.uf.edex.esb.camel/META-INF/MANIFEST.MF
	viz/com.raytheon.uf.viz.personalities.cave/splash.bmp
	viz/com.raytheon.viz.core.gl/META-INF/MANIFEST.MF
	viz/com.raytheon.viz.core.gl/src/com/raytheon/viz/core/gl/glsl/GLShaderProgram.java
	viz/com.raytheon.viz.core.gl/src/com/raytheon/viz/core/gl/internal/GLTarget.java
	viz/com.raytheon.viz.ui/src/com/raytheon/viz/ui/dialogs/ImagingDialog.java

+++ common/com.raytheon.uf.common.derivparam.python/utility/common_static/base/derivedParameters/functions/TW.py
@@ -1,22 +1,31 @@
 ##
 # This software was developed and / or modified by Raytheon Company,
 # pursuant to Contract DG133W-05-CQ-1067 with the US Government.
-# 
+#
 # U.S. EXPORT CONTROLLED TECHNICAL DATA
 # This software product contains export-restricted data whose
 # export/transfer/disclosure is restricted by U.S. law. Dissemination
 # to non-U.S. persons whether in the United States or abroad requires
 # an export license or other authorization.
-# 
+#
 # Contractor Name:        Raytheon Company
 # Contractor Address:     6825 Pine Street, Suite 340
 #                         Mail Stop B8
 #                         Omaha, NE 68106
 #                         402.291.0100
-# 
+#
 # See the AWIPS II Master Rights File ("Master Rights File.pdf") for
 # further licensing information.
 ##
+
+#
+# SOFTWARE HISTORY
+#
+# Date           Ticket#      Engineer      Description
+# ------------   ----------   -----------   -----------
+#                             ????          Initial creation
+# Aug 05, 2015   4703         njensen       cast -999999.0 to float32
+#
 
 ## @file TW.py
 import gridslice
@@ -39,7 +48,7 @@ c1 = 0.0091379024
 ## @var c2: constant used in multiple functions.
 c2 = 6106.3960
 
-## 
+##
 # Calculate wet bulb temperature in degrees C
 # from pressure, temperature, and relative humidity.
 #
@@ -49,7 +58,7 @@ c2 = 6106.3960
 # @return: wet bulb temperature in degrees C as a numpy array 
 #
 def execute1(P, T, RH):
-    
+
     # Find dry-bulb temp from temp and relative humidity
     RH = clip(RH,1.0,100.0)
     b = T * c1
@@ -59,7 +68,7 @@ def execute1(P, T, RH):
     val -= 223.1986
     td = b - sqrt(val)
     td /= 0.0182758048
-    
+
     TW = myTW(T, td, P)
     return TW
 
@@ -74,9 +83,9 @@ def execute2(GH3D, T3D, RH3D, TILT):
     vc3d = GH3D[0]
     slice3d = T3D[0]
     slice = gridslice.createNumpySlice(vc3d, slice3d, TILT, 2)
-    slice = where(slice == 9.99999993e+36, - 999999.0, slice)
+    slice = where(slice == 9.99999993e+36, float32(-999999.0), slice)
     return slice
-    
+
 ##
 # This function takes arrays of values and filters them against expected values
 # for terrestrial weather. Where the dewpoint temperature is greater than the
@@ -89,7 +98,7 @@ def execute2(GH3D, T3D, RH3D, TILT):
 # @param Td: Calculated dewpoint temperature in degrees Kelvin.
 # @param P: Pressure in millibars
 # @return: TW, the wet-bulb temperature
-# @rtype: assumed to be a numpy array. 
+# @rtype: assumed to be a numpy array.
 def myTW(T, Td, P):
     td_lt_t = Td < T
     t_ge_cold = T>=100.0
@@ -98,13 +107,13 @@ def myTW(T, Td, P):
 
     satVP = zeros(T.shape, dtype=T.dtype)
     satVP[typical] = calcSatVP(T[typical])
-    
+
     norm_svp = satVP <= 10.0
     sane = typical & norm_svp
 
     veryCold = td_lt_t & ~t_ge_cold
     td_satvp_bad = ~( td_lt_t & norm_svp ) # includes NaN cells
-    
+
     if rank(P)==0:
         sel = lambda val, mask : val
     else:
